i386/readelflib.c: add support for ARM libraries
We include ARM's readelflib.c and add some code to the ELF parser, so
it'll hand off parsing ARM & Aarch64 files to the ARM routine.

The resulting ldconfig, which runs on x86, is, thus, able to parse ARM
libraries properly and generate a correct ld.so.cache. This can be used
during cross-compilation to generate ld.so.cache for ARM on the x86
build host.

Signed-off-by: Markus Mayer <mmayer@broadcom.com>